REFACTOR: Cut out the fat
