--- lib/cloudformation-ruby-dsl.rb
@@ -1,2 +1,0 @@
-require "cloudformation-ruby-dsl/version"
-
